Use ts-node  and revert to CommonJS Fixes #1206
