+++ .mocharc.cjs
@@ -1,3 +1,7 @@
 module.exports = {
-  "enable-source-maps": true
+  "enable-source-maps": true,
+  "extension": ["ts"],
+  "spec": "packages/**/*.test.ts",
+  "require": "ts-node/register",
+  "recursive": true
 }
