+++ packages/mcos-auth/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-auth",
-    "type": "module",
     "exports": "./built/index.js"
 }
