+++ packages/mcos-compression/package.json
@@ -1,5 +1,4 @@
 {
   "name": "mcos-compression",
-  "type": "module",
-  "exports": "./built/index.js"
+  "exports": "./src/index.js"
 }
