+++ packages/mcos-compression/src/testHelpers.js
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { ReadBuff, WriteBuff } from '../src/types.js'
+import { ReadBuff, WriteBuff } from './types.js'
 
 /**
  * @description
