+++ packages/mcos-compression/tsconfig.json
@@ -4,7 +4,7 @@
     "module": "ES2022",
     "outDir": "./built",
     "rootDirs": ["./src", "./test"],
-    "baseUrl": "."
+    "baseUrl": ".",
   },
   "exclude": [
     "built"
