+++ packages/mcos-database/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-database",
-    "type": "module",
     "exports": "./built/index.js"
 }
