+++ packages/mcos-gateway/src/adminServer.test.ts
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import chai from 'chai'
-import { AdminServer } from 'mcos-gateway'
+import * as chai from 'chai'
+import { AdminServer } from './adminServer.js'
 import { describe, it } from 'mocha'
 
 chai.should()
