+++ packages/mcos-lobby/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-lobby",
-    "type": "module",
     "exports": "./built/index.js"
 }
