+++ packages/mcos-login/package.json
@@ -1,6 +1,5 @@
 {
     "name": "mcos-login",
-    "type": "module",
     "exports": {
       ".": "./built/index.js"
     }
