+++ packages/mcos-patch/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-patch",
-    "type": "module",
     "exports": "./built/index.js"
 }
