+++ packages/mcos-persona/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-persona",
-    "type": "module",
     "exports": "./built/index.js"
 }
