+++ packages/mcos-shared/src/structures/GSMessageBase.test.ts
@@ -14,8 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import chai from 'chai'
-import { GSMessageBase } from 'mcos-shared/structures'
+import * as chai from 'chai'
+import { GSMessageBase } from './GMessageBase.js'
 import { describe, it } from 'mocha'
 
 chai.should()
