+++ packages/mcos-transactions/package.json
@@ -1,5 +1,4 @@
 {
     "name": "mcos-transactions",
-    "type": "module",
     "exports": "./built/index.js"
 }
